fix: made it easier to reuse, fixed some width errors.
