+++ src/screens/Home.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import Button from '../components/styled/Button';
 import { ComponentVariantType } from '../utils/constants';
 import { ButtonLabelSize } from '../components/styled/Button/styles';
@@ -7,8 +7,29 @@ import { ProgramItem } from '../components/program/ProgramItem';
 import { RootContainer, StyledColumn, StyledRow } from '../components/styled/styles';
 import ProgramsList from '../components/home/ProgramsList';
 import ProfessorList from '../components/home/ProfessorList';
+import { Dropdown } from '../components/Dropdown';
 
 const Home = () => {
+  const [selectedValue, setSelectedValue] = useState<string>('');
+  const content = [
+    'Option 1',
+    'Option 2',
+    'Oliver Hansen',
+    'Van Henry',
+    'April Tucker',
+    'Ralph Hubbard',
+    'Omar Alexander',
+    'Carlos Abbott',
+    'Miriam Wagner',
+    'Bradley Wilkerson',
+    'Virginia Andrews',
+    'Kelly Snyder',
+  ];
+
+  const handleChange = (value: string) => {
+    setSelectedValue(value);
+  };
+
   const handleButton = () => {
     alert('hello');
   };
@@ -25,7 +46,8 @@ const Home = () => {
     >
       <StyledRow css={{ width: '1100px', height: '870px', gap: '40px' }}>
         <ProgramsList />
-        <StyledColumn css={{ flex: 1 }}>
+        <StyledColumn css={{ flex: 1, gap: '12px' }}>
+          <Dropdown onChange={handleChange} value={selectedValue} content={content} />
           <ProfessorList />
         </StyledColumn>
       </StyledRow>
